# openapi.yaml
openapi: "3.0.3"
info:
  title: War Tiket Konser API
  description: |
    Sistem penjualan tiket konser dengan antrean virtual dan kursi terbatas.
    Kelompok 5 — Praktikum Microservices.
    
    **Masalah utama:** Satu kursi hanya boleh terjual satu kali walau seribu orang memintanya bersamaan.
    
    **Services:**
    - event-service (port 3001) — Kelola konser, jadwal, kategori kursi, dan harga
    - ticket-service (port 3002) — Kunci kursi sementara, konfirmasi, lepas kunci saat kedaluwarsa
    - payment-service (port 3003) — Terima pembayaran, konfirmasi, batalkan bila gagal/timeout
    - notification-service (port 3004) — Kirim e-ticket, pengingat, dan pemberitahuan gagal bayar
  version: "1.0.0"
  contact:
    name: Kelompok 5
servers:
  - url: http://localhost:3001
    description: event-service
  - url: http://localhost:3002
    description: ticket-service
  - url: http://localhost:3003
    description: payment-service
  - url: http://localhost:3004
    description: notification-service
...
paths:
...
  /catalog:
    get:
      tags: [catalog]
      summary: Daftar konser aktif dengan kursi tersedia
      description: |
        **ENDPOINT KRITIS** — Endpoint ini diakses ratusan ribu kali saat tiket dibuka.
        Menampilkan semua konser yang masih memiliki kursi tersedia beserta harga dan jumlahnya.
      operationId: getCatalog
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Daftar konser berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/EventWithSeats"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
  /events:
    get:
      tags: [events]
      summary: Semua konser (termasuk yang sudah habis)
      operationId: getEvents
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
    post:
      tags: [events]
      summary: Buat konser baru (admin)
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Konser berhasil dibuat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          $ref: "#/components/responses/BadRequest"
  /events/{eventId}:
    get:
      tags: [events]
      summary: Detail satu konser
      operationId: getEventById
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventWithSeats"
        "404":
          $ref: "#/components/responses/NotFound"
  /events/{eventId}/seats:
    get:
      tags: [catalog]
      summary: Kursi tersedia untuk satu konser
      operationId: getEventSeats
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  event_id:
                    type: integer
                  categories:
                    type: array
                    items:
                      $ref: "#/components/schemas/SeatCategory"

  # ══════════════════════════════════════════════════════
  # TICKET-SERVICE (port 3002)
  # ══════════════════════════════════════════════════════
...
  /payments:
    post:
      tags: [payments]
      summary: Proses pembayaran untuk pesanan yang terkunci
      description: |
        **ENDPOINT KRITIS** — Menerima pembayaran untuk order yang sudah ada.
        Jika pembayaran berhasil, mengirim event ke RabbitMQ untuk konfirmasi tiket dan notifikasi.
        Jika gagal atau timeout, kunci kursi dilepas otomatis.
      operationId: createPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
      responses:
        "201":
          description: Pembayaran berhasil diproses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Order sudah kedaluwarsa atau sudah dibayar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /payments/{paymentId}:
    get:
      tags: [payments]
      summary: Status pembayaran
      operationId: getPaymentById
      parameters:
        - name: paymentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          $ref: "#/components/responses/NotFound"
  /payments/{paymentId}/cancel:
    post:
      tags: [payments]
      summary: Batalkan pembayaran (refund)
      operationId: cancelPayment
      parameters:
        - name: paymentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Pembayaran dibatalkan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ══════════════════════════════════════════════════════
  # NOTIFICATION-SERVICE (port 3004)
  # ══════════════════════════════════════════════════════
...
components:
  parameters:
    eventId:
      name: eventId
      in: path
      required: true
      schema: { type: integer }
...
  responses:
    NotFound:
      description: Data tidak ditemukan
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Request tidak valid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Event:
      type: object
      properties:
        id:          { type: integer, example: 1 }
        name:        { type: string, example: "Konser Dewa 19 Reuni" }
        venue:       { type: string, example: "Gelora Bung Karno, Jakarta" }
        event_date:  { type: string, format: date-time }
        description: { type: string }
        banner_url:  { type: string }
        status:
          type: string
          enum: [upcoming, on_sale, sold_out, completed, cancelled]
        created_at:  { type: string, format: date-time }
    EventWithSeats:
      allOf:
        - $ref: "#/components/schemas/Event"
        - type: object
          properties:
            categories:
              type: array
              items:
                $ref: "#/components/schemas/SeatCategory"
    SeatCategory:
      type: object
      properties:
        id:              { type: integer }
        event_id:        { type: integer }
        name:            { type: string, example: "VIP" }
        total_seats:     { type: integer, example: 500 }
        available_seats: { type: integer, example: 247 }
        price:           { type: number, format: double, example: 750000 }
    CreateEventRequest:
      type: object
      required: [name, venue, event_date]
      properties:
        name:        { type: string }
        venue:       { type: string }
        event_date:  { type: string, format: date-time }
        description: { type: string }
        banner_url:  { type: string }
        categories:
          type: array
          items:
            type: object
            required: [name, total_seats, price]
            properties:
              name:        { type: string }
              total_seats: { type: integer }
              price:       { type: number }
...
    Payment:
      type: object
      properties:
        id:             { type: integer }
        order_id:       { type: integer }
        user_id:        { type: string }
        amount:         { type: number }
        method:
          type: string
          enum: [bank_transfer, credit_card, gopay, ovo, dana]
        status:
          type: string
          enum: [pending, success, failed, cancelled, refunded]
        paid_at:        { type: string, format: date-time, nullable: true }
        created_at:     { type: string, format: date-time }
    CreatePaymentRequest:
      type: object
      required: [order_id, user_id, method]
      properties:
        order_id: { type: integer, example: 42 }
        user_id:  { type: string, example: "user_105841117023" }
        method:
          type: string
          enum: [bank_transfer, credit_card, gopay, ovo, dana]
          example: gopay
...
    Error:
      type: object
      properties:
        error:   { type: string }
        message: { type: string }
        code:    { type: integer }